+++ internal/processor/processorfx/module.go
@@ -2,7 +2,7 @@ package processorfx
 
 import (
 	"github.com/bitmagnet-io/bitmagnet/internal/processor"
-	"github.com/bitmagnet-io/bitmagnet/internal/processor/queue/handler"
+	"github.com/bitmagnet-io/bitmagnet/internal/processor/queue"
 	"go.uber.org/fx"
 )
 
@@ -11,7 +11,7 @@ func New() fx.Option {
 		"processor",
 		fx.Provide(
 			processor.New,
-			handler.New,
+			queue.New,
 		),
 	)
 }
