+++ internal/processor/queue/handler.go
@@ -1,4 +1,4 @@
-package handler
+package queue
 
 import (
 	"context"
